# docs/commercialization/openapi/open-v1.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "YINYU CTF Platform Open API",
    "description": "YINYU CTF Platform external API contract",
    "version": "v1"
  },
  "paths": {
    "/api/open/v1/teamlab/runtimes": {
      "post": {
        "tags": [
          "TeamLab - Runtimes"
        ],
        "summary": "Create a runtime",
        "description": "Queues deployment of a released topology for one team or automation owner.",
        "operationId": "OpenTeamLabRuntimes_Create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "x-originalName": "idempotencyKey",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamLabRuntimeModel"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
    "/api/open/v1/teamlab/runtimes/{runtimeId}": {
      "get": {
        "tags": [
          "TeamLab - Runtimes"
        ],
        "summary": "Get a runtime",
        "description": "Returns the aggregated runtime, shard, network, and asset state.",
        "operationId": "OpenTeamLabRuntimes_Get",
        "parameters": [
          {
            "name": "runtimeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenTeamLabRuntimeModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "TeamLab - Runtimes"
        ],
        "summary": "Destroy a runtime",
        "description": "Queues cleanup of all runtime shards, assets, routes, captures, and access grants.",
        "operationId": "OpenTeamLabRuntimes_Destroy",
        "parameters": [
          {
            "name": "runtimeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "Idempotency-Key",
            "x-originalName": "idempotencyKey",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
    "/api/open/v1/teamlab/runtimes/{runtimeId}/reset": {
      "post": {
        "tags": [
          "TeamLab - Runtimes"
        ],
        "summary": "Reset a runtime",
        "description": "Queues a controlled cleanup and redeployment using the runtime release and optional overlays.",
        "operationId": "OpenTeamLabRuntimes_Reset",
        "parameters": [
          {
            "name": "runtimeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "Idempotency-Key",
            "x-originalName": "idempotencyKey",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetTeamLabRuntimeModel"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
    "/api/open/v1/teamlab/runtimes/{runtimeId}/events": {
      "get": {
        "tags": [
          "TeamLab - Runtimes"
        ],
        "summary": "List runtime events",
        "description": "Returns cursor-paginated lifecycle and deployment events for troubleshooting and audit.",
        "operationId": "OpenTeamLabRuntimes_Events",
        "parameters": [
          {
            "name": "runtimeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 100.0,
              "minimum": 1.0
            },
            "x-position": 3
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenTeamLabRuntimeEventPageModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
    "/api/open/v1/teamlab/runtimes/{runtimeId}/access-grants": {
      "post": {
        "tags": [
          "TeamLab - Runtimes"
        ],
        "summary": "Create a WireGuard access grant",
        "description": "Queues creation of a short-lived, single-download player access configuration.",
        "operationId": "OpenTeamLabRuntimes_CreateAccessGrant",
        "parameters": [
          {
            "name": "runtimeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "Idempotency-Key",
            "x-originalName": "idempotencyKey",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamLabAccessGrantCreateModel"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
    "/api/open/v1/teamlab/runtimes/{runtimeId}/access-grants/{grantId}/download": {
      "get": {
        "tags": [
          "TeamLab - Runtimes"
        ],
        "summary": "Download an access configuration",
        "description": "Consumes the one-time download token and returns the WireGuard configuration file.",
        "operationId": "OpenTeamLabRuntimes_DownloadAccessConfiguration",
        "parameters": [
          {
            "name": "runtimeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "grantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          },
          {
            "name": "token",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "400": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalApiProblemDetailsModel"
                }
              }
            }
          },
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    },
...
    "/api/open/v1/operations/{id}": {
      "get": {
        "tags": [
          "Operations"
        ],
        "operationId": "Operations_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiOperationModel"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "GzctfApiToken": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ExternalApiProblemDetailsModel": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProblemDetails"
          },
          {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "properties": {
              "code": {
                "type": "string"
              },
              "traceId": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "CreateTeamLabRuntimeModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "releaseId": {
            "type": "string",
            "format": "guid"
          },
          "externalReference": {
            "type": "string",
            "nullable": true
          },
          "constraints": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TeamLabRuntimeConstraintsModel"
              }
            ]
          },
          "overlays": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TeamLabRuntimeOverlayModel"
            }
          }
        }
      },
...
      "ResetTeamLabRuntimeModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "overlays": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TeamLabRuntimeOverlayModel"
            }
          },
          "releaseId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          }
        }
      },
      "OpenTeamLabRuntimeEventPageModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamLabRuntimeEventModel"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "GzctfApiToken": {
        "type": "http",
        "description": "Scoped GZCTF API token",
        "scheme": "bearer",
        "bearerFormat": "GZCTF API token"
      }
    }
  },
...
}